requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
SQLAlchemy>=2.0.25
python-dotenv>=1.0.1
requests>=2.31.0
pandas>=2.2.2
numpy>=1.26.4
streamlit>=1.35.0
python-dateutil>=2.9.0.post0
altair>=5.3.0
